re-envision role system: introduce role customization, make frontend react to granular permissions (#129)
* role-revamp: introduce role customization, make frontend react to permissions instead of role

* bruh...

* code cleanup

* teacher-flow: introduce perms for lp-delete, lr-delete

* code cleanup

* fix: regional-scope filter overridable by client query params

* fix: wrong permission string on regenerate endpoint

* fix: getRolePermissions() does not filter-out deleted roles

* role: fix isSystem being used in places where isSuperuser was expected
The logic is, isSuperuser makes the role immutable while isSystem only makes the role "not deletable".

* role: introduce discriminator for scoped flows

* code cleanup

* code cleanup

* backend: add integration tests

* backend: drop integration test via cicd

* code cleanup

* fix: session leaks server-side encrypted otp, recovry, rememberMeToken

* user: reinstate admin profile tab

* setProfile: only update if parent teacher profile exists

* migration: retain db in consistent state on midtask failure

* auth: version jwts

* bug fix: drop isDeleted in PUT /user/:id

* make profileType mandatory, validate before permission selection

* bug fix: reject edits to role held by actorr

* bug fix: interrupted-run recovery races other replicas

* make audit references non-blocking, keep strict checks for operational scalar refs

* role: gate AI lesson edits by permission

* test: align unit mocks with current request contracts

* migration: seed roles on fresh databases

* migration: version unify-users to prevent reruns

* fix: integrate scoped roles with Superset dashboards

* post-rebase changes

+++ shiksha-website/shiksha-frontend/src/app/view/user/profile/profile.component.ts
@@ -14,7 +14,7 @@ import { UtilityService } from 'src/app/core/services/utility.service';
 import { MasterService } from 'src/app/shared/services/master.service';
 import { Router } from '@angular/router';
 import { SidebarService } from 'src/app/layout/sidebar/sidebar.service';
-import { forkJoin, of } from 'rxjs';
+import { forkJoin } from 'rxjs';
 import { TranslateService } from '@ngx-translate/core';
 
 @Component({
@@ -23,7 +23,6 @@ import { TranslateService } from '@ngx-translate/core';
   styleUrls: ['./profile.component.scss'],
 })
 export class ProfileComponent implements OnInit, OnDestroy {
-  isTeacherRole!: boolean;
   showDeleteClassDetailsConfirm!: boolean;
   showDeleteResourceConfirm!: boolean;
   selectedClassIndex!:any;
@@ -117,6 +116,7 @@ export class ProfileComponent implements OnInit, OnDestroy {
   userData: any;
   userPorfileForm!: FormGroup;
   loggedInUser: any;
+  isTeacher!: boolean;
   dependentPatchData: any;
   resourceMasterData: any;
   boardMasterData: any;
@@ -134,8 +134,6 @@ export class ProfileComponent implements OnInit, OnDestroy {
   patchObj: any;
 
   defaultBoard = null;
-
-  defaultMedium = null;
 
   showDeleteProfileImageConfirm =false;
 
@@ -157,7 +155,7 @@ export class ProfileComponent implements OnInit, OnDestroy {
   ngOnInit(): void {
     const data: string = localStorage.getItem('userData') ?? '';
     this.loggedInUser = JSON.parse(data);
-    this.isTeacherRole = this.utilityService.hasPermission(['standard', 'power']);
+    this.isTeacher = Boolean(this.loggedInUser.profiles.teacher);
     this.createUserForm();
     this.getData();
   }
@@ -166,17 +164,17 @@ export class ProfileComponent implements OnInit, OnDestroy {
    * Function to get all master data and profile data
    */
   getData() {
-    const boardMaster = this.isTeacherRole
-      ? this.service.getClassesByBoard(this.loggedInUser.school._id)
-      : of(null);
-    const resourceMaster = this.masterService.getFacilities();
     const userProfile = this.service.getProfileInfo(this.loggedInUser._id);
+    if (!this.isTeacher) {
+      userProfile.subscribe({ next: (profile) => this.setProfileInfo(profile), error: (err) => this.utilityService.handleError(err) });
+      return;
+    }
 
+    const boardMaster = this.service.getClassesByBoard(this.loggedInUser.school._id);
+    const resourceMaster = this.masterService.getFacilities();
     forkJoin([boardMaster, resourceMaster, userProfile]).subscribe({
       next: ([boardRes, resourceRes, profileRes]) => {
-        if (this.isTeacherRole) {
-          this.setClassesByBoard(boardRes);
-        }
+        this.setClassesByBoard(boardRes);
         this.setResourceData(resourceRes);
         this.setProfileInfo(profileRes);
       },
@@ -199,9 +197,6 @@ export class ProfileComponent implements OnInit, OnDestroy {
   presetValues(data: any) {
     if (data.length === 1) {
       this.defaultBoard = data[0]._id;
-      if (data[0].medium.length === 1) {
-        this.defaultMedium = data[0].medium[0].medium;
-      }
     }
   }
 
@@ -210,15 +205,16 @@ export class ProfileComponent implements OnInit, OnDestroy {
    * @param i
    * @param val
    */
-  setMediumSubjectDropdown(i: any, val: any, mode: any) {
+  setMediumSubjectDropdown(i: any, val: any) {
     this.resetclassInfo('board', i);
     if (val) {
       this.mediumDropdownOptions[i] = val.medium;
       this.currentSubjects[i] = val.subjects;
       // this.subjectDropdownOptions[i] = val.subjects;
-      if (val.medium.length === 1 && mode === 'add') {
+      if (val.medium.length === 1) {
         this.classes.controls[i].get('medium')?.setValue(val.medium[0].medium);
-        this.classDropdownOptions[i] = val.medium[0].classDetails;
+        this.classes.controls[i].get('medium')?.disable();
+        this.setClassDropdown(i, val.medium[0]);
       }
     }
   }
@@ -227,6 +223,11 @@ export class ProfileComponent implements OnInit, OnDestroy {
     this.resetclassInfo('medium', i);
     if (val) {
       this.classDropdownOptions[i] = val.classDetails;
+      if (val.classDetails.length === 1) {
+        this.classes.controls[i].get('class')?.setValue(val.classDetails[0].standard);
+        this.classes.controls[i].get('class')?.disable();
+        this.setStrength(i, val.classDetails[0]);
+      }
     }
   }
 
@@ -240,6 +241,12 @@ export class ProfileComponent implements OnInit, OnDestroy {
 
     if (val) {
       this.subjectDropdownOptions[i] = this.filterSubjects(val.standard,this.currentSubjects[i])
+      if (this.subjectDropdownOptions[i].length === 1) {
+        const subject = this.subjectDropdownOptions[i][0];
+        this.classes.controls[i].get('subject')?.setValue(subject._id);
+        this.classes.controls[i].get('subject')?.disable();
+        this.subjectMapper(i, subject);
+      }
       this.resetclassInfo('strength', i);
       this.classes.controls[i].get('boysStrength')?.setValue(val.boysStrength);
       this.classes.controls[i]
@@ -257,23 +264,23 @@ export class ProfileComponent implements OnInit, OnDestroy {
    */
   resetclassInfo(type: any, i: any) {
     if (type === 'board') {
-      this.classes.controls[i].get('medium')?.reset();
-      this.classes.controls[i].get('class')?.reset();
-      this.classes.controls[i].get('subject')?.reset();
+      this.classes.controls[i].get('medium')?.reset({ value: null, disabled: false });
+      this.classes.controls[i].get('class')?.reset({ value: null, disabled: false });
+      this.classes.controls[i].get('subject')?.reset({ value: null, disabled: false });
       this.classes.controls[i].get('subjectDetails')?.reset();
       this.classDropdownOptions[i] = [];
       this.mediumDropdownOptions[i] = [];
       this.subjectDropdownOptions[i] = [];
     } else if (type === 'medium') {
-      this.classes.controls[i].get('class')?.reset();
-      this.classes.controls[i].get('subject')?.reset();
+      this.classes.controls[i].get('class')?.reset({ value: null, disabled: false });
+      this.classes.controls[i].get('subject')?.reset({ value: null, disabled: false });
       this.classes.controls[i].get('subjectDetails')?.reset();
       this.classDropdownOptions[i] = [];
       this.subjectDropdownOptions[i] = [];
     }
     else if (type === 'standard') {
       this.subjectDropdownOptions[i] = [];
-      this.classes.controls[i].get('subject')?.reset();
+      this.classes.controls[i].get('subject')?.reset({ value: null, disabled: false });
       this.classes.controls[i].get('subjectDetails')?.reset();
     }
     this.classes.controls[i].get('boysStrength')?.reset();
@@ -284,15 +291,17 @@ export class ProfileComponent implements OnInit, OnDestroy {
    * Function to get profile info
    */
   setProfileInfo(val:any) {
-    if (this.isTeacherRole) {
-      this.mergeSchoolResource(val?.data?.school?.facilities);
-    }
+    this.userData = val.data;
+    if (!this.isTeacher) return;
 
-    this.userData = val?.data;
+    const teacherProfile = val?.data?.profiles?.teacher;
+    const schoolFacilities = val.data.school.facilities;
+    this.mergeSchoolResource(schoolFacilities);
+
     const keysToRemove = ['classes', 'facilities'];
 
     const { newObj, removedObj } = this.utilityService.removeKeys(
-      val?.data,
+      teacherProfile,
       keysToRemove
     );
     this.patchObj = newObj;
@@ -335,9 +344,8 @@ export class ProfileComponent implements OnInit, OnDestroy {
     this.userPorfileForm = this.fb.group({
       classes: this.fb.array([]),
       facilities: this.fb.array([]),
-      preferredLanguage: [this.loggedInUser?.preferredLanguage],
     });
-    this.addResource();
+    if (this.isTeacher) this.addResource();
   }
 
   /**
@@ -404,6 +412,12 @@ export class ProfileComponent implements OnInit, OnDestroy {
   setSubjectDropdown(i:any,val:any,standard:any){
     if (val) {
       this.subjectDropdownOptions[i] = this.filterSubjects(standard,val.subjects);
+      if (this.subjectDropdownOptions[i].length === 1) {
+        const subject = this.subjectDropdownOptions[i][0];
+        this.classes.controls[i].get('subject')?.setValue(subject._id);
+        this.classes.controls[i].get('subject')?.disable();
+        this.subjectMapper(i, subject);
+      }
     }
   }
   /**
@@ -414,10 +428,13 @@ export class ProfileComponent implements OnInit, OnDestroy {
       this.classes.controls[i]
         ?.get('board')
         ?.setValue(this.dependentPatchData.classes[i].board);
+      if (this.boardDropdownOptions.length === 1) {
+        this.classes.controls[i].get('board')?.disable();
+      }
       const mediums = this.boardMasterData.filter(
         (e: any) => e._id === this.dependentPatchData.classes[i].board
       );
-      this.setMediumSubjectDropdown(i, mediums[0], 'edit');
+      this.setMediumSubjectDropdown(i, mediums[0]);
       this.classes.controls[i]
         ?.get('medium')
         ?.setValue(this.dependentPatchData.classes[i].medium);
@@ -515,7 +532,6 @@ export class ProfileComponent implements OnInit, OnDestroy {
       subscribe({
       next:(res)=>{
           this.loggedInUser.preferredLanguage = lang;
-          this.userPorfileForm.get('preferredLanguage')?.setValue(lang);
           this.utilityService.handleResponse(res);
           localStorage.setItem('userData', JSON.stringify(this.loggedInUser));
           this.translateService.use(lang);
@@ -594,22 +610,11 @@ export class ProfileComponent implements OnInit, OnDestroy {
    */
   defaultPreset() {
     if (this.defaultBoard) {
+      const i = this.classes.length - 1;
       this.boardDropdownOptions = this.boardMasterData;
-      this.mediumDropdownOptions[this.classes.length - 1] =
-        this.boardMasterData[0].medium;
-      // this.subjectDropdownOptions[this.classes.length - 1] =
-      //   this.boardMasterData[0].subjects;
-      this.currentSubjects[this.classes.length - 1] = this.boardMasterData[0].subjects;
-      this.classes.controls[this.classes.length - 1]
-        .get('board')
-        ?.setValue(this.defaultBoard);
-      if (this.defaultMedium) {
-        this.classDropdownOptions[this.classes.length - 1] =
-          this.boardMasterData[0].medium[0].classDetails;
-        this.classes.controls[this.classes.length - 1]
-          .get('medium')
-          ?.setValue(this.defaultMedium);
-      }
+      this.classes.controls[i].get('board')?.setValue(this.defaultBoard);
+      this.classes.controls[i].get('board')?.disable();
+      this.setMediumSubjectDropdown(i, this.boardMasterData[0]);
     }
   }
 
@@ -740,14 +745,14 @@ export class ProfileComponent implements OnInit, OnDestroy {
       return;
     }
 
-    const classDetails = this.splitClasses(this.classes.value);
+    const classDetails = this.splitClasses(this.classes.getRawValue());
 
     if(this.utilityService.hasDuplicates(classDetails)){
       this.utilityService.showWarning('Duplicate class-subject mapping found. Please verify.');
       return
     }
 
-    const data = this.userPorfileForm.value;
+    const data = this.userPorfileForm.getRawValue();
     data.facilities = this.utilityService.removeObjectsWithEmptyType(
       data.facilities
     );
@@ -757,7 +762,8 @@ export class ProfileComponent implements OnInit, OnDestroy {
     this.service.updateProfile(data).subscribe({
       next: (res) => {
         this.utilityService.handleResponse(res);
-        localStorage.setItem('userData', JSON.stringify(res.data));
+        this.loggedInUser.profiles.teacher = res.data.profiles.teacher;
+        localStorage.setItem('userData', JSON.stringify(this.loggedInUser));
         this.router.navigate(['/']);
       },
       error: (err) => {
